pyproject.toml
[tool.poetry]
name = "pdf-parser"
version = "0.1.0"
description = "A Python package for parsing and processing PDF documents"
authors = ["Oxford Data Processes"]
readme = "README.md"
packages = [{include = "pdf_parser"}]

[tool.poetry.dependencies]
python = "^3.8"
pdfplumber = "^0.10.2"
pytest = "^7.4.3"
pandas = "^2.1.3"
pathlib = "^1.0.1"
pytesseract = "^0.3.10"
Pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.27.1"
jsonschema = "^4.20.0"
fitz = "^0.0.1.dev2"
pydantic = "^2.5.2"
matplotlib = "^3.8.2"
PyMuPDF = "^1.23.7"
mypy = "^1.7.1"
types-Pillow = "^10.1.0.2"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
